fix: refactor priority alerts
