+++ package.json
@@ -58,7 +58,7 @@
     "tailwind-scrollbar-hide": "^1.1.7",
     "tailwindcss": "^3.1.8",
     "tailwindcss-scoped-groups": "^2.0.0",
-    "typescript": "5.7.3",
+    "typescript": "5.8.2",
     "uuid": "^11.0.1",
     "viem": "2.x",
     "wagmi": "^2.10.10",
@@ -87,7 +87,7 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "22.13.5",
+    "@types/node": "22.13.9",
     "@typescript-eslint/eslint-plugin": "^7.1.0",
     "cross-spawn": "^7.0.6",
     "eslint": "8.57.0",
